refactor(backend): extract trace crud into crud/telemetry.py
Move the OpenTelemetry trace/span CRUD out of the crud/__init__.py monolith
into a dedicated crud/telemetry.py, following the existing crud/explorer.py
split. Pure move: every function body is byte-identical to the original.

The block was self-contained - it called no other function in crud/__init__.py
and used none of the crud_utils helpers - so the extraction creates no import
cycle. Four schema imports (OTELSpanCreate, StatusCode, TraceSource, TraceType)
leave crud/__init__.py entirely since nothing else there used them.

Names are re-exported from crud/__init__.py so existing callers keep working;
the next commit migrates them and drops the shim.
